plop
git-svn-id: http://svn.net-core.org/repos/t-engine4@1455 51575b47-30f0-44d4-a5cc-537603b46e54

+++ game/modules/boot/data/gfx/shaders/main_fbo.frag
@@ -68,7 +68,7 @@ void main(void)
 				}
 			}
 		}
-		sample /= float((motionblur*1.5) * ((motionblur*0.5));
+		sample /= (motionblur*1.5) * (motionblur*0.5);
 //		gl_FragColor = sample * (0.3 + noise * 0.7);
 		gl_FragColor = sample;
 	}
